# backend/docs/swagger.yaml
basePath: /api
definitions:
  dtos.RegisterRequest:
    properties:
      email:
        type: string
      fullName:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  dtos.RegisterResponse:
    properties:
      email:
        type: string
      message:
        type: string
      userId:
        type: string
      username:
        type: string
    type: object
  handlers.ChangePasswordRequest:
    properties:
      currentPassword:
        type: string
      newPassword:
        type: string
    type: object
...
  handlers.ForgotPasswordRequest:
    properties:
      email:
        type: string
    type: object
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handlers.LoginResponse:
    properties:
      token:
        description: 'Optional: If we want to issue token from backend directly'
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  handlers.ResendVerificationRequest:
    properties:
      email:
        type: string
    type: object
  handlers.ResetPasswordRequest:
    properties:
      newPassword:
        type: string
      token:
        type: string
    type: object
...
  models.Permission:
    properties:
      action:
        description: e.g., 'create', 'read', 'update', 'delete'
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        description: e.g., 'query:create'
        type: string
      resource:
        description: e.g., 'query', 'dashboard'
        type: string
    type: object
...
  models.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_system_role:
        description: TRUE for built-in roles
        type: boolean
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: string
    type: object
...
  models.User:
    properties:
      createdAt:
        type: string
      deactivatedAt:
        type: string
      deactivatedBy:
        type: string
      deactivationReason:
        type: string
      email:
        type: string
      emailVerified:
        type: boolean
      emailVerifiedAt:
        type: string
      id:
        type: string
      name:
        type: string
      provider:
        description: OAuth fields
        type: string
      providerId:
        description: OAuth provider user ID
        type: string
      role:
        description: user, admin
        type: string
      roles:
        description: RBAC
        items:
          $ref: '#/definitions/models.Role'
        type: array
      status:
        description: Status
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@insightengine.ai
    name: API Support
  description: Enterprise BI Platform API with Power BI/Tableau parity.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: InsightEngine API
  version: "1.0"
paths:
...
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: Changes password after verifying current password. Requires authentication.
      parameters:
      - description: Change Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - Auth
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Generates reset token and sends reset email. Always returns success.
      parameters:
      - description: Forgot Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Request password reset
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates user and returns JWT token. Checks if email is verified.
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: User login
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with validated input and sends a verification
        email.
      parameters:
      - description: Registration Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - Auth
  /auth/resend-verification:
    post:
      consumes:
      - application/json
      description: Resends verification email to user.
      parameters:
      - description: Resend Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ResendVerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Resend verification email
      tags:
      - Auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Resets user password using valid token.
      parameters:
      - description: Reset Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Reset password
      tags:
      - Auth
  /auth/verify-email:
    get:
      description: Verifies user email using token from email link.
      parameters:
      - description: Verification Token
        in: query
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Verify email address
      tags:
      - Auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"